refactor(ui): make SSH key components reusable
Make the SSH key list and form reusable and migrate the settings table to TypeScript.

Fixes: canonical-web-and-design/app-squad#163.
