+++ ui/src/app/base/components/SSHKeyForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SSHKeyForm";
