+++ ui/src/app/base/components/SSHKeyList/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SSHKeyList";
