+++ ui/src/app/preferences/views/SSHKeys/AddSSHKey/AddSSHKey.test.tsx
@@ -1,5 +1,4 @@
 import { mount } from "enzyme";
-import { act } from "react-dom/test-utils";
 import { Provider } from "react-redux";
 import { MemoryRouter } from "react-router-dom";
 import configureStore from "redux-mock-store";
@@ -39,25 +38,6 @@ describe("AddSSHKey", () => {
     expect(wrapper.find("AddSSHKey").exists()).toBe(true);
   });
 
-  it("cleans up when unmounting", async () => {
-    const store = mockStore(state);
-    const wrapper = mount(
-      <Provider store={store}>
-        <MemoryRouter initialEntries={["/"]}>
-          <AddSSHKey />
-        </MemoryRouter>
-      </Provider>
-    );
-
-    act(() => {
-      wrapper.unmount();
-    });
-
-    expect(
-      store.getActions().some((action) => action.type === "sshkey/cleanup")
-    ).toBe(true);
-  });
-
   it("redirects when the SSH key is saved", () => {
     state.sshkey.saved = true;
     const store = mockStore(state);
@@ -69,84 +49,5 @@ describe("AddSSHKey", () => {
       </Provider>
     );
     expect(wrapper.find("Redirect").exists()).toBe(true);
-  });
-
-  it("can upload an SSH key", () => {
-    const store = mockStore(state);
-    const wrapper = mount(
-      <Provider store={store}>
-        <MemoryRouter initialEntries={["/"]}>
-          <AddSSHKey />
-        </MemoryRouter>
-      </Provider>
-    );
-    act(() =>
-      wrapper.find("Formik").props().onSubmit({
-        key: "ssh-rsa...",
-      })
-    );
-    expect(
-      store.getActions().find((action) => action.type === "sshkey/create")
-    ).toStrictEqual({
-      type: "sshkey/create",
-      payload: {
-        params: {
-          key: "ssh-rsa...",
-        },
-      },
-      meta: {
-        model: "sshkey",
-        method: "create",
-      },
-    });
-  });
-
-  it("can import an SSH key", () => {
-    const store = mockStore(state);
-    const wrapper = mount(
-      <Provider store={store}>
-        <MemoryRouter initialEntries={["/"]}>
-          <AddSSHKey />
-        </MemoryRouter>
-      </Provider>
-    );
-    act(() =>
-      wrapper.find("Formik").props().onSubmit({
-        auth_id: "wallaroo",
-        protocol: "lp",
-      })
-    );
-    expect(
-      store.getActions().find((action) => action.type === "sshkey/import")
-    ).toStrictEqual({
-      type: "sshkey/import",
-      payload: {
-        params: {
-          auth_id: "wallaroo",
-          protocol: "lp",
-        },
-      },
-      meta: {
-        model: "sshkey",
-        method: "import_keys",
-      },
-    });
-  });
-
-  it("adds a message when a SSH key is added", () => {
-    state.sshkey.saved = true;
-    const store = mockStore(state);
-    mount(
-      <Provider store={store}>
-        <MemoryRouter initialEntries={["/"]}>
-          <AddSSHKey />
-        </MemoryRouter>
-      </Provider>
-    );
-    const actions = store.getActions();
-    expect(actions.some((action) => action.type === "sshkey/cleanup")).toBe(
-      true
-    );
-    expect(actions.some((action) => action.type === "message/add")).toBe(true);
   });
 });
