+++ ui/src/app/preferences/views/SSHKeys/AddSSHKey/AddSSHKey.tsx
@@ -1,67 +1,25 @@
-import { useDispatch, useSelector } from "react-redux";
 import { useHistory } from "react-router-dom";
-import * as Yup from "yup";
-
-import SSHKeyFormFields from "../SSHKeyFormFields";
-import type { SSHKeyFormValues } from "../SSHKeyFormFields/types";
 
 import FormCard from "app/base/components/FormCard";
-import FormikForm from "app/base/components/FormikForm";
-import { useAddMessage, useWindowTitle } from "app/base/hooks";
+import SSHKeyForm from "app/base/components/SSHKeyForm";
+import { useWindowTitle } from "app/base/hooks";
 import prefsURLs from "app/preferences/urls";
-import { actions as sshkeyActions } from "app/store/sshkey";
-import sshkeySelectors from "app/store/sshkey/selectors";
-
-const SSHKeySchema = Yup.object().shape({
-  protocol: Yup.string().required("Source is required"),
-  auth_id: Yup.string().when("protocol", {
-    is: (val: string) => val && val !== "upload",
-    then: Yup.string().required("ID is required"),
-  }),
-  key: Yup.string().when("protocol", {
-    is: (val: string) => val === "upload",
-    then: Yup.string().required("Key is required"),
-  }),
-});
 
 export const AddSSHKey = (): JSX.Element => {
-  const dispatch = useDispatch();
   const history = useHistory();
-  const errors = useSelector(sshkeySelectors.errors);
-  const saved = useSelector(sshkeySelectors.saved);
-  const saving = useSelector(sshkeySelectors.saving);
-
   useWindowTitle("Add SSH key");
-
-  useAddMessage(saved, sshkeyActions.cleanup, "SSH key successfully imported.");
 
   return (
     <FormCard title="Add SSH key">
-      <FormikForm<SSHKeyFormValues>
-        cleanup={sshkeyActions.cleanup}
-        errors={errors}
-        initialValues={{ auth_id: "", protocol: "", key: "" }}
+      <SSHKeyForm
         onCancel={() => history.push({ pathname: prefsURLs.sshKeys.index })}
         onSaveAnalytics={{
           action: "Saved",
           category: "SSH keys preferences",
           label: "Import SSH key form",
         }}
-        onSubmit={(values) => {
-          if (values.key && values.key !== "") {
-            dispatch(sshkeyActions.create(values));
-          } else {
-            dispatch(sshkeyActions.import(values));
-          }
-        }}
-        saving={saving}
-        saved={saved}
         savedRedirect={prefsURLs.sshKeys.index}
-        submitLabel="Import SSH key"
-        validationSchema={SSHKeySchema}
-      >
-        <SSHKeyFormFields />
-      </FormikForm>
+      />
     </FormCard>
   );
 };
