+++ ui/src/scss/index.scss
@@ -77,6 +77,7 @@
 @import "~app/base/components/Placeholder";
 @import "~app/base/components/Popover";
 @import "~app/base/components/Section";
+@import "~app/base/components/SSHKeyList";
 @import "~app/base/components/TableMenu";
 @import "~app/base/components/TagSelector";
 @include ColumnToggle;
@@ -89,6 +90,7 @@
 @include Placeholder;
 @include Popover;
 @include Section;
+@include SSHKeyList;
 @include TableMenu;
 @include TagSelector;
 
@@ -194,11 +196,9 @@
 
 // preferences
 @import "~app/preferences/views/APIKeys/APIKeyList";
-@import "~app/preferences/views/SSHKeys/SSHKeyList";
 @import "~app/preferences/views/SSLKeys/AddSSLKey";
 @import "~app/preferences/views/SSLKeys/SSLKeyList";
 @include APIKeyList;
-@include SSHKeyList;
 @include AddSSLKey;
 @include SSLKeyList;
